Quash context (#12)
* Rename initial_cwd() to activeDirectory().

* Update makefile to include path.h & quashcontext.h

* Add ability to print all environment variables.

* Add contextCtor()

Allocates memory for and creates a QuashContext struct. A pointer
to this memory is returned. It is the responsibility of the callee
to deallocate. This function also sets the current working directory.

* Add contextDtor()

This function will serve as a Destructor of sorts for the
QuashContext struct. Pass it a struct QuashContext* and
contextDtor() will free all of it's dynamically allocated
members and then the struct itself as well.

* Add ability to update QC CWD.

* Fix contextDtor().

* Fix split().

split() from quashutils.h was incorrectly removing final characters
from strings since the final NULL character was not properly accounted
for in the size(third) parameter of strlcpy().

* Validate that QuashContext ready for testing.

QuashContext _should_ be prepared for testing with the `cd`
command. The hope is that our qc-cwd will be abkle to be updated
by calling qc->updateCWD().

* Fix strlcpy() size allocation issue.